% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{aggregation.apply}
\alias{aggregation.apply}
\title{apply aggregation function}
\usage{
aggregation.apply(groupby, aggfun, aggcol, dt)
}
\arguments{
\item{groupby}{the list of columns name that will be grouped}

\item{aggfun}{the function to aggregate with}

\item{aggcol}{the list of columns name that will be aggregated}

\item{dt}{the input dataframe}
}
\value{
a dataframe
}
\description{
do aggregation and return a new data table.
the groupby columns must include problem and algorithm
}
\examples{
aggregation.apply(groupby= c("problem", "algorithm"), aggfun= c("mean"), aggcol= c("measure.mmce.test.mean", "measure.ber.test.mean"), dt= mlr.benchmark.example)
}
