% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{batch.benchmark}
\alias{batch.benchmark}
\title{Batch benchmark}
\description{
A example of a benchmark created with mlr and batchtools.
The table consists of 11 columns and 12 rows.
Problems: iris, sonar, mlbench.waveform, mlbench.ringnorm
Algorithms: lda, rpart, randomForest
Replications: CV 2.
The file is accessible through "batch.benchmark".
}
\references{
See \url{data/batch.benchmark.RData}
}
\keyword{data}
