% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{check.aggregation.valid}
\alias{check.aggregation.valid}
\title{check if the input function is a valid aggregation function}
\usage{
check.aggregation.valid(agg.fun)
}
\arguments{
\item{agg.fun}{aggregation function}
}
\value{
boolean
}
\description{
check if the input function is a valid aggregation function (i.e., returns a single numeric value)
}
