% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{check.transform.list.to.value}
\alias{check.transform.list.to.value}
\title{check if the input function is a valid list-to-value transformation function}
\usage{
check.transform.list.to.value(transform.fun)
}
\arguments{
\item{transform.fun}{transformation function}
}
\value{
boolean
}
\description{
check if the input function is a valid list-to-value transformation function
(i.e., applied to a list of lists of numeric values
returns a list of numeric values of the same length)
[transformation R^n x R^m => R^n]
}
