% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{check.transform.valid}
\alias{check.transform.valid}
\title{check if the input function is a valid transformation function}
\usage{
check.transform.valid(transform.fun)
}
\arguments{
\item{transform.fun}{transformation function}
}
\value{
boolean
}
\description{
check if the input function is a valid transformation function (i.e., applied to a list
of numeric values returns a list of numeric values of the same length)
}
