% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStructure.R
\name{checkColumnNames}
\alias{checkColumnNames}
\title{Check the columns of a benchmarkVis compatible data table}
\usage{
checkColumnNames(dt)
}
\arguments{
\item{dt}{data table to check}
}
\value{
True if test was successful
}
\description{
Check if data table just contains legit columns.
Should be the six standard columns (problem, problem.parameter, algorithm, ...) + measures + lists
}
\examples{
checkColumnNames(microbenchmark.example)
}
