% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{column.type}
\alias{column.type}
\title{get type of values of a column (numeric values, vector or other)}
\usage{
column.type(column)
}
\arguments{
\item{column}{column of a data frame to check}
}
\value{
string description of values' type
}
\description{
check if a data frame's column contains numeric values, 
vectors or is of other type
}
