% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{createBarPlot}
\alias{createBarPlot}
\title{Create a bar plot}
\usage{
createBarPlot(dt, measure, group.problems = FALSE)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}

\item{group.problems}{if true group results by problem (default: FALSE)}
}
\value{
a plotly bar plot
}
\description{
Create a plotly bar plot out of a benchmarkVis compatible data table.
The created bar plot shows the measure result for each benchmark entry.
x-Axis: the measure.
y-Axis: the benchmark entries.
}
\examples{
createBarPlot(mlr.benchmark.example, "measure.mmce.test.mean", FALSE)
}
