% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxPlot.R
\name{createBoxPlot}
\alias{createBoxPlot}
\title{Create a box plot}
\usage{
createBoxPlot(dt, measure, violin = FALSE, facetting_problem = FALSE)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{measure for comparison}

\item{violin}{if set to TRUE a violin plot instead of boxplot is produced (facetting_problem should be FALSE)}

\item{facetting_problem}{if set to TRUE boxplot with facet wrap is produced (violin should be FALSE)}
}
\value{
a box plot
}
\description{
Create a plotly box plot out of a benchmarkVis compatible data table.
The created box plot allows for comparison of algorithms based on a given performance measure for different datasets
}
\examples{
createBoxPlot(mlr.benchmark.example, 'measure.mmce.test.mean')
createBoxPlot(mlr.benchmark.example, 'measure.mmce.test.mean', violin = TRUE)
createBoxPlot(mlr.benchmark.example, 'measure.mmce.test.mean', facetting_problem = TRUE)
}
