% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critDifferencesPlot.R
\name{createCritDifferencesPlot}
\alias{createCritDifferencesPlot}
\title{create a crit differences plot}
\usage{
createCritDifferencesPlot(dt, measure, group.by = "problem",
  test.str = "bd", p.value = 0.05)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the measure to plot}

\item{group.by}{the column to group the markers by. Possibilities: "algorithm", "problem", "replication" (default: "problem")}

\item{test.str}{Test for which the critical differences are computed. “bd” for the Bonferroni-Dunn Test. “nemenyi” for the [PMCMR::posthoc.friedman.nemenyi.test] which is comparing all classifiers to each other. Possibilities: "nemenyi", "bd"}

\item{p.value}{P-value for the critical difference}
}
\value{
a crit differences plot
}
\description{
Create critDifferences plot out of a benchmarkVis compatible data table.
}
\examples{
createCritDifferencesPlot(mlr.benchmark.example, "measure.mmce.test.mean")
}
