% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityPlot.R
\name{createDensityPlot}
\alias{createDensityPlot}
\title{Create a density plot}
\usage{
createDensityPlot(dt, measure, stack.plots = FALSE)
}
\arguments{
\item{dt}{campatible data table}

\item{measure}{the column name of the measure}

\item{stack.plots}{defines if the density curves should be stacked. Alternative is transparent. Default: FALSE}
}
\value{
a plotly density plot
}
\description{
Create a plotly density plot out of a benchmarkVis compatible data table.
The created density chart shows the distribution of the specified measure for each algorithm
}
\examples{
createDensityPlot(mlr.benchmark.example, "measure.mmce.test.mean", TRUE)
}
