% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityRankPlot.R
\name{createDensityRankPlot}
\alias{createDensityRankPlot}
\title{Create a density rank plot}
\usage{
createDensityRankPlot(dt, ignore.measures = vector(), stack.plots = FALSE,
  color.by = "algorithm", group.by = "problem")
}
\arguments{
\item{dt}{compatible data table}

\item{ignore.measures}{a vector of strings describing which measures to leave out of the plot (default: empty)}

\item{stack.plots}{defines if the density curves should be stacked. Alternative is transparent. Default: FALSE}

\item{color.by}{the column to color the input with. Possibilities: "algorithm", "problem", "replication" (default: "algorithm")}

\item{group.by}{the column to group the ranks by. Possibilities: "algorithm", "problem", "replication" (default: "problem")}
}
\value{
a plotly density rank plot
}
\description{
Create a plotly sensity plot out of a benchmarkVis compatible data table.
The created plot shows the distribution of the ranks each input achieves depending on a group and measures.
x-Axis: the ranks.
y-Axis: the density.
}
\examples{
createDensityRankPlot(mlr.benchmark.example, c("measure.timetrain.test.mean","measure.mmce.test.mean"))
}
