% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualMeasurePlot.R
\name{createDualMeasurePlot}
\alias{createDualMeasurePlot}
\title{Create a Dual Measure plot}
\usage{
createDualMeasurePlot(dt, measure1, measure2, color.algorithm.only = TRUE,
  pointsize = 4L, jitter = 0)
}
\arguments{
\item{dt}{compatible data table}

\item{measure1}{the first measure for comparison}

\item{measure2}{the second measure  for comparison}

\item{color.algorithm.only}{if set to true, each color corresponds to an algorithm, otherwise to algorithm and problem}

\item{pointsize}{the size of ploted point}

\item{jitter}{small vertical jitter to deal with overplotting in case of equal scores}
}
\value{
a dual measure plot
}
\description{
Create a Dual Measure plot out of a benchmarkVis compatible data table.
The created Dual Measure plot allows to compare the performance of interaction(algorithm and problem) based on two giving performance measures.
x-Axis: measure1.
y-Axis: measure2.
color: the interaction of algorithm and problem.
}
\examples{
createDualMeasurePlot(mlr.benchmark.example, 'measure.mmce.test.mean','measure.ber.test.mean')
}
