% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationBoxPlot}
\alias{createIterationBoxPlot}
\title{Create a iteration box plot}
\usage{
createIterationBoxPlot(dt, measure, violin = FALSE,
  iteration.algorithm = "default")
}
\arguments{
\item{dt}{campatible data table}

\item{measure}{the measure on the y axis}

\item{violin}{if set to TRUE a violin plot instead of boxplot is produced (default: FALSE)}

\item{iteration.algorithm}{the algorithm to investigate. Algorithm.parameter must contain "iteration" field.
(default: "default" - would take the first from getIterationAlgorithms())}
}
\value{
a plotly box plot
}
\description{
Create a plotly box plot out of a benchmarkVis compatible data table.
The created box chart shows the distribution for the input measure.
}
\examples{
createIterationBoxPlot(mlr.tuning.example, "measure.acc.test.mean")
}
