% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationDensityPlot}
\alias{createIterationDensityPlot}
\title{Create a iteration density plot}
\usage{
createIterationDensityPlot(dt, measure, iteration.algorithm = "default")
}
\arguments{
\item{dt}{campatible data table}

\item{measure}{the input measure on the x axis}

\item{iteration.algorithm}{the algorithm to investigate. Algorithm.parameter must contain "iteration" field.
(default: "default" - would take the first from getIterationAlgorithms())}
}
\value{
a plotly density plot
}
\description{
Create a plotly density plot out of a benchmarkVis compatible data table.
The created density chart shows the distribution of the specified measure.
}
\examples{
createIterationDensityPlot(mlr.tuning.example, "measure.acc.test.mean")
}
