% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationDualMeasurePlot}
\alias{createIterationDualMeasurePlot}
\title{Create a iteration line plot with a measure on x and y aixs}
\usage{
createIterationDualMeasurePlot(dt, measure1, measure2, draw.lines = FALSE,
  regression.line = FALSE, iteration.algorithm = "default")
}
\arguments{
\item{dt}{campatible data table}

\item{measure1}{the measure on the x axis}

\item{measure2}{the measure on the y axis}

\item{draw.lines}{draw a line between the points in the original order (default: FALSE)}

\item{regression.line}{add a regression line to the plot (default: FALSE)}

\item{iteration.algorithm}{the algorithm to investigate. Algorithm.parameter must contain "iteration" field.
(default: "default" - would take the first from getIterationAlgorithms())}
}
\value{
a plotly line plot
}
\description{
Create a plotly line plot out of a benchmarkVis compatible data table.
The created line chart shows the change within the specified list measures.
}
\examples{
createIterationDualMeasurePlot(mlr.tuning.example, "measure.acc.test.mean", "measure.acc.test.sd")
}
