% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationDualParameterPlot}
\alias{createIterationDualParameterPlot}
\title{Create a iteration dual parameter scatter plot}
\usage{
createIterationDualParameterPlot(dt, measure, parameter,
  parameter2 = "iteration")
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}

\item{parameter}{the first parameter you want to examine}

\item{parameter2}{the second parameter you want to examine (default: "iteration")}
}
\value{
a plotly scatter plot
}
\description{
This plot is just usable if the algorithm.parameter column contains an iteration value.
Create a plotly scatter to plot out of a benchmarkVis compatible data table.
The created scatter plot shows the connection between the specified measure and two different algorithm parameters.
x-Axis: the second parameter - default is iteration.
y-Axis: the first parameter.
Size and color of the circles: Depends on the measure value.
}
\examples{
createIterationDualParameterPlot(mlr.tuning.example, "measure.acc.test.mean", "C", "sigma")
}
