% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationMeasureMatrixPlot}
\alias{createIterationMeasureMatrixPlot}
\title{Create a iteration measure matrix plot}
\usage{
createIterationMeasureMatrixPlot(dt, iteration.algorithm = "default")
}
\arguments{
\item{dt}{campatible data table}

\item{iteration.algorithm}{the algorithm to investigate. Algorithm.parameter must contain "iteration" field.
(default: "default" - would take the first from getIterationAlgorithms())}
}
\value{
a plotly measure matrix plot
}
\description{
Create a plotly measure matrix plot out of a benchmarkVis compatible data table.
The created measure matrix chart shows the distribution of each measure as density and scatter plot.
}
\examples{
createIterationMeasureMatrixPlot(mlr.tuning.example)
}
