% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationParameterPlot}
\alias{createIterationParameterPlot}
\title{Create a iteration parameter scatter plot}
\usage{
createIterationParameterPlot(dt, measure, parameter, show.histogram = FALSE,
  iteration.algorithm = "default")
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}

\item{parameter}{the algorithm parameter you want to examine}

\item{show.histogram}{shows the histogram of the measure values in the background (default: FALSE)}

\item{iteration.algorithm}{the algorithm to investigate. Algorithm.parameter must contain "iteration" field.
(default: "default" - would take the first from getIterationAlgorithms())}
}
\value{
a plotly scatter plot
}
\description{
This plot is just usable if the algorithm.parameter column contains an iteration value.
Create a plotly scatter plot out of a benchmarkVis compatible data table.
The created scatter plot shows the connection between the specified measure and algorithm parameter.
x-Axis: the algorithm parameter.
y-Axis: the measure.
}
\examples{
createIterationParameterPlot(mlr.tuning.example, "measure.acc.test.mean", "C", TRUE)
}
