% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterationPlots.R
\name{createIterationScatterPlot}
\alias{createIterationScatterPlot}
\title{Create a iteration scatter plot}
\usage{
createIterationScatterPlot(dt, measure, iteration.algorithm = "default")
}
\arguments{
\item{dt}{campatible data table}

\item{measure}{the measure on the y axis}

\item{iteration.algorithm}{the algorithm to investigate. Algorithm.parameter must contain "iteration" field.
(default: "default" - would take the first from getIterationAlgorithms())}
}
\value{
a plotly scatter plot
}
\description{
Create a plotly scatter plot out of a benchmarkVis compatible data table.
The created scatter chart shows the different values for the measure.
}
\examples{
createIterationScatterPlot(mlr.tuning.example, "measure.acc.test.mean")
}
