% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListBoxPlot}
\alias{createListBoxPlot}
\title{Create a list box plot}
\usage{
createListBoxPlot(dt, list.measure, violin = FALSE)
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure}{the column name containing the list of a specific measure}

\item{violin}{if set to TRUE a violin plot instead of boxplot is produced (default: FALSE)}
}
\value{
a plotly box plot
}
\description{
Create a plotly box plot out of a benchmarkVis compatible data table.
The created box plot allows for comparison of the benchmark entries based on a given performance measure
x-Axis: the benchmark entries.
y-Axis: measure.
}
\examples{
createListBoxPlot(microbenchmark.example, "list.values")
}
