% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListDensityRankPlot}
\alias{createListDensityRankPlot}
\title{Create a list density rank plot}
\usage{
createListDensityRankPlot(dt, list.measure, stack.plots = FALSE)
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure}{the column name containing the list of a specific measure}

\item{stack.plots}{defines if the density curves should be stacked. Alternative is transparent. Default: FALSE}
}
\value{
a plotly density plot
}
\description{
Create a plotly density plot out of a benchmarkVis compatible data table.
The created density chart shows the distribution of the specified list measure by its ranks.
}
\examples{
createListDensityRankPlot(microbenchmark.example, "list.values", TRUE)
}
