% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListDualMeasurePlot}
\alias{createListDualMeasurePlot}
\title{Create a list line plot with a measure on x and y aixs}
\usage{
createListDualMeasurePlot(dt, list.measure1, list.measure2,
  draw.lines = FALSE, color.by = "algorithm", style.by = "problem")
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure1}{the list measure on the x axis}

\item{list.measure2}{the list measure on the y axis}

\item{draw.lines}{draw a line between the points in the original order (default: FALSE)}

\item{color.by}{defines the color of the markers. Possibilities: "algorithm", "problem", "replication" (default: "algorithm")}

\item{style.by}{defines the style of the line. Just working with 'draw.lines' = TRUE. Possibilities: "algorithm", "problem", "replication" (default: "problem")}
}
\value{
a plotly line plot
}
\description{
Create a plotly line plot out of a benchmarkVis compatible data table.
The created line chart shows the change within the specified list measures
}
\examples{
createListDualMeasurePlot(mlr.benchmark.example, "list.mmce", "list.ber")
}
