% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListMeasureMatrixPlot}
\alias{createListMeasureMatrixPlot}
\title{Create a list measure matrix plot}
\usage{
createListMeasureMatrixPlot(dt)
}
\arguments{
\item{dt}{compatible data table}
}
\value{
a measure matrix plot
}
\description{
Create a list measure matrix plot out of a benchmarkVis compatible data table.
The created matrix plot shows all measure results for each benchmark entry.
BEWARE! All list measures need to have the same length.
}
\examples{
createListMeasureMatrixPlot(mlr.benchmark.example)
}
