% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListRankMatrixBarPlot}
\alias{createListRankMatrixBarPlot}
\title{Create a list rank matrix bar plot}
\usage{
createListRankMatrixBarPlot(dt, list.measure, stacked = TRUE)
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure}{the column name containing the list of a specific measure}

\item{stacked}{defines if the bars should be stacked (default: TRUE)}
}
\value{
a plotly rank matrix bar plot
}
\description{
Create a plotly rank matrix bar plot out of a benchmarkVis compatible data table.
The created bar chart shows the the rank of the specified list measure for each iteration.
}
\examples{
createListRankMatrixBarPlot(microbenchmark.example, "list.values")
}
