% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListScatterPlot}
\alias{createListScatterPlot}
\title{Create a list scatter plot}
\usage{
createListScatterPlot(dt, list.measure)
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure}{the column name containing the list of a specific measure}
}
\value{
a plotly scatter plot
}
\description{
Create a plotly scatter plot out of a benchmarkVis compatible data table.
The created scatter plot allows to compare the performances based on the giving performance measure.
x-Axis: the benchmark entries.
y-Axis: measure.
}
\examples{
createListScatterPlot(microbenchmark.example, "list.values")
}
