% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/measureMatrixPlot.R
\name{createMeasureMatrixPlot}
\alias{createMeasureMatrixPlot}
\title{Create a measure matrix plot}
\usage{
createMeasureMatrixPlot(dt, color.by = "algorithm")
}
\arguments{
\item{dt}{compatible data table}

\item{color.by}{the column to color the markers with. Possibilities: "algorithm", "problem", "replication" (default: "algorithm")}
}
\value{
a measure matrix plot
}
\description{
Create a measure matrix plot out of a benchmarkVis compatible data table.
The created matrix plot shows all measure results for each benchmark entry.
}
\examples{
createMeasureMatrixPlot(mlr.benchmark.example)
}
