% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterPlots.R
\name{createParameterDualPlot}
\alias{createParameterDualPlot}
\title{Create a parameter dual scatter plot}
\usage{
createParameterDualPlot(dt, measure, parameter.column, parameter,
  parameter.column2, parameter2, plot.area = FALSE)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}

\item{parameter.column}{the parameter column which contains the parameter}

\item{parameter}{the first parameter you want to examine}

\item{parameter.column2}{the parameter column which contains the second parameter}

\item{parameter2}{the second parameter you want to examine}

\item{plot.area}{plot the measure as an area instead of scatters (default: FALSE)}
}
\value{
a plotly scatter plot
}
\description{
This plot is just usable if at least one parameter column is contained.
Create a plotly scatter plot out of a benchmarkVis compatible data table.
The created scatter plot shows the connection between the specified measure and two different parameters.
x-Axis: the second parameter.
y-Axis: the first parameter.
Size and color of the circles: Depends on the measure value.
}
\examples{
createParameterDualPlot(mlr.tuning.example, "measure.acc.test.mean", "algorithm.parameter", "C", "algorithm.parameter", "sigma")
}
