% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterPlots.R
\name{createParameterMeasureMatrixPlot}
\alias{createParameterMeasureMatrixPlot}
\title{Create a parameter measure matrix plot}
\usage{
createParameterMeasureMatrixPlot(dt, parameter.column, parameter)
}
\arguments{
\item{dt}{campatible data table}

\item{parameter.column}{the parameter column which contains the parameter}

\item{parameter}{the parameter you want to examine}
}
\value{
a plotly measure matrix plot
}
\description{
This plot is just usable if at least one parameter column is contained in the data table.
Create a plotly measure matrix plot out of a benchmarkVis compatible data table.
The created measure matrix chart shows the distribution of each measure as density and scatter plot.
}
\examples{
createParameterMeasureMatrixPlot(mlr.tuning.example, "algorithm.parameter", "C")
}
