% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameterPlots.R
\name{createParameterScatterPlot}
\alias{createParameterScatterPlot}
\title{Create a parameter scatter plot}
\usage{
createParameterScatterPlot(dt, measure, parameter.column, parameter,
  color.by = "algorithm", show.histogram = FALSE, regression.line = FALSE)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}

\item{parameter.column}{the parameter column which contains the parameter}

\item{parameter}{the parameter you want to examine}

\item{color.by}{the column to color the markers with. Possibilities: "algorithm", "problem", "replication" (default: "algorithm")}

\item{show.histogram}{shows the histogram of the measure values in the background (default: FALSE)}

\item{regression.line}{add a regression line to the plot (default: FALSE)}
}
\value{
a plotly scatter plot
}
\description{
This plot is just usable if at least one parameter column is contained in the data table.
Create a plotly scatter plot out of a benchmarkVis compatible data table.
The created scatter plot shows the connection between the specified measure and parameter values.
x-Axis: the parameter.
y-Axis: the measure.
}
\examples{
createParameterScatterPlot(mlr.tuning.example, "measure.acc.test.mean", "algorithm.parameter", "C", "algorithm", TRUE)
}
