% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radarPlot.R
\name{createRadarPlot}
\alias{createRadarPlot}
\title{Create a radar plot}
\usage{
createRadarPlot(dt)
}
\arguments{
\item{dt}{compatible data table}
}
\value{
a radar plot
}
\description{
Create a radar plot out of a benchmarkVis compatible data table.
The created radar plot shows all measure results for each benchmark entry.
}
\examples{
createRadarPlot(microbenchmark.example)
}
