% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{createRankMatrixBarPlot}
\alias{createRankMatrixBarPlot}
\title{Create a rank matrix bar plot}
\usage{
createRankMatrixBarPlot(dt, ignore.measures = vector())
}
\arguments{
\item{dt}{compatible data table}

\item{ignore.measures}{a vector of strings describing which measures to leave out of the plot (default: empty)}
}
\value{
a plotly rank matrix bar plot
}
\description{
Create a plotly rank matrix bar plot out of a benchmarkVis compatible data table.
The created bar plot shows the frequency of the ranks each algorithm achieves depending on problem and measure.
x-Axis: the ranks.
y-Axis: the frequency.
}
\examples{
createRankMatrixBarPlot(mlr.benchmark.example, c("measure.timetrain.test.mean","measure.mmce.test.mean"))
}
