% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicationPlots.R
\name{createReplicationLinePlot}
\alias{createReplicationLinePlot}
\title{Create a replication line plot}
\usage{
createReplicationLinePlot(dt, list.measure)
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure}{the column name containing the replications of a specific measure}
}
\value{
a plotly line plot
}
\description{
Create a plotly line plot out of a benchmarkVis compatible data table.
The created line chart shows the change within the specified measure for each iteration of the replication
}
\examples{
createReplicationLinePlot(microbenchmark.example, "list.values")
}
