% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningPlots.R
\name{createTuningIterationPlot}
\alias{createTuningIterationPlot}
\title{Create a tuning iteration line plot}
\usage{
createTuningIterationPlot(dt, measure)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}
}
\value{
a plotly line plot
}
\description{
Create a plotly line plot out of a benchmarkVis compatible data table.
The created line plot shows the change within the specified measure for each iteration
}
\examples{
createTuningIterationPlot(mlr.tuning.example, "measure.acc.test.mean")
}
