% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuningPlots.R
\name{createTuningParameterPlot}
\alias{createTuningParameterPlot}
\title{Create a tuning parameter scatter plot}
\usage{
createTuningParameterPlot(df, parameter, measure)
}
\arguments{
\item{df}{campatible dataframe}

\item{parameter}{the parameter you want to examine}

\item{measure}{the column name containing the results of a measure}
}
\value{
a plotly scatter plot
}
\description{
Create a plotly scatter plot out of a benchmarkVis compatible dataframe.
The created scatter plot shows the change within the specified measure for a change of the specified parameter
}
\examples{
createTuningParameterPlot(mlr.tuning.example, "C", "acc.test.mean")
}
