% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAggregation.R
\name{do.agg}
\alias{do.agg}
\title{do aggregation result}
\usage{
\method{do}{agg}(fun, groupby, aggcol, df)
}
\arguments{
\item{fun}{aggregate functions}

\item{groupby}{the list of columns name that will be grouped}

\item{aggcol}{the list of columns name that will be aggregated}

\item{df}{the input dataframe}
}
\value{
a dataframe
}
\description{
do aggregation and return a new data table.
the aggcol must include problem and algorithm.
only nummeric columns can be aggregated.
}
