% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{eval.function}
\alias{eval.function}
\title{evaluate the input function}
\usage{
eval.function(fun)
}
\arguments{
\item{fun}{input function (string or function)}
}
\value{
function
}
\description{
if the input is a string parses and evaluates the function contained
}
