% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{Example}
\alias{Example}
\title{This file is a basic Version of an useable benchmark result dataframe.
The file can be loaded with "readRDS("./data/Example.rda")"}
\description{
This file is a basic Version of an useable benchmark result dataframe.
The file can be loaded with "readRDS("./data/Example.rda")"
}
\references{
\url{data/Example.rda}
}
\keyword{data}
