% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAggregation.R
\name{get.agg.result}
\alias{get.agg.result}
\title{get aggregation result}
\usage{
get.agg.result(fun_str, groupby, aggcol, df)
}
\arguments{
\item{fun_str}{the string of the aggregate functions}

\item{groupby}{the list of columns name that will be grouped}

\item{aggcol}{the list of columns name that will be aggregated}

\item{df}{the input dataframe}
}
\value{
a dataframe
}
\description{
do aggregation and return a new data table
}
