% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{get.num.columns.name}
\alias{get.num.columns.name}
\title{get numeric columns' names for a dataframe}
\usage{
get.num.columns.name(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
list of the numeric column names
}
\description{
detects numeric columns' names for a dataframe
for numeric values as well as lists of numeric values
}
