% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAggregation.R
\name{get.result}
\alias{get.result}
\title{get aggregation result}
\usage{
get.result(groupby, aggfun, aggcol, df)
}
\arguments{
\item{groupby}{the list of columns name that will be grouped}

\item{aggcol}{the list of columns name that will be aggregated}

\item{df}{the input dataframe}
}
\value{
a dataframe
}
\description{
do aggregation and return a new data table.
the groupby columns must include problem and algorithm
}
\examples{
get.result(groupby= c("problem", "algorithm"), aggfun= c("mean"), aggcol= c("mmce.test.mean", "ber.test.mean"), df= mlr.benchmark.example)
}
