% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getIterationAlgorithms}
\alias{getIterationAlgorithms}
\title{Get all iteration algorithms}
\usage{
getIterationAlgorithms(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
vector containing all iteration algorithm names
}
\description{
Get the names of algorithms with multiple iterations within the data table.
Algorithms for which the parameters contain "iteration" field.
}
\examples{
getIterationAlgorithms(mlr.tuning.example)
}
