% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getLists}
\alias{getLists}
\title{Get all lists}
\usage{
getLists(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
vector containing all list names
}
\description{
Get the names of list columns within the data table.
List columns start with "list."
}
\examples{
getLists(mlr.benchmark.example)
}
