% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getMainColumns}
\alias{getMainColumns}
\title{Get all main columns}
\usage{
getMainColumns(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
vector containing all main column names
}
\description{
Get the names of main columns within the data table.
Can be "problem", "algorithm", "replication". "Problem" and "Algorithm" are mandatory.
}
\examples{
getMainColumns(mlr.benchmark.example)
}
