% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getMeasures}
\alias{getMeasures}
\title{Get all measures}
\usage{
getMeasures(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
vector containing all measure names
}
\description{
Get the names of measure columns within the data table.
Measures columns start with "measure."
}
\examples{
getMeasures(mlr.benchmark.example)
}
