% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getParameterColumns}
\alias{getParameterColumns}
\title{Get all parameter columns}
\usage{
getParameterColumns(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
vector containing all parameter column names
}
\description{
Get the names of parameter columns within the data table.
Can be "problem.parameter", "algorithm.parameter", "replication.parameter".
}
\examples{
getParameterColumns(mlr.benchmark.example)
}
