% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getParameters}
\alias{getParameters}
\title{Get all parameters of a specified parameter column}
\usage{
getParameters(dt, parameter.column)
}
\arguments{
\item{dt}{input data}

\item{parameter.column}{the parameter column (can be "problem.parameter", "algorithm.parameter", "replication.parameter")}
}
\value{
vector containing all parameter names for the parameter column
}
\description{
Get the names of parameters in the specified parameter column within the data table.
}
\examples{
getParameters(mlr.benchmark.example, "algorithm.parameter")
}
