% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPrettyListMeasureName}
\alias{getPrettyListMeasureName}
\title{Get name of a list measure without "list." prefix}
\usage{
getPrettyListMeasureName(measure.name)
}
\arguments{
\item{measure.name}{list measure name}
}
\value{
name of the list measure without "list."
}
\description{
Get name of a list measure without "list." prefix
}
