% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPrettyMeasureName}
\alias{getPrettyMeasureName}
\title{Get name of a measure without "measure." prefix}
\usage{
getPrettyMeasureName(measure.name)
}
\arguments{
\item{measure.name}{measure name}
}
\value{
name of the measure without "measure."
}
\description{
Get name of a measure without "measure." prefix
}
