% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getPrettyPlotList}
\alias{getPrettyPlotList}
\title{Get list of all possible plots in a readable format: "Type: Name"}
\usage{
getPrettyPlotList(input.list)
}
\arguments{
\item{input.list}{the plot list you want to beautify}
}
\value{
vector containing all plots in a readable format
}
\description{
Get a complete list of all plots implemented within the benchmarkVis application. Uses the listPlots() method and transforms the result into the better readable format: "Type: Name"
}
