% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getPrettyPlotName}
\alias{getPrettyPlotName}
\title{Get name of a plot in a readable format: "Type: Name"}
\usage{
getPrettyPlotName(plot.name)
}
\arguments{
\item{plot.name}{the plot you want to beautify}
}
\value{
name of the plot in a readable format
}
\description{
Get a prettified name of a plot implemented within the benchmarkVis application.
}
