% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getTunings}
\alias{getTunings}
\title{Get all tuning sessions}
\usage{
getTunings(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
list containing vectors of form: (problem, algorithm, replicaiton)
}
\description{
Get the combination of (problem, algorithm, replicaiton) with multiple iterations within the data table.
Should be used to tune the algorithm parameters.
Combinations can be recognised by the occurrence of the "iteration" field in the algorithm.parameter column.
}
\examples{
getTunings(mlr.tuning.example)
}
