% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{getValidPlots}
\alias{getValidPlots}
\title{Get list of all plots useable with the data table}
\usage{
getValidPlots(dt)
}
\arguments{
\item{dt}{input data table}
}
\value{
Return a list of plots suitable for the given data
}
\description{
Get a complete list of all plots implemented within the benchmarkVis application which are combinable with the input data table.
(depends on the presence/absence of measures, iteration and list data)
}
\examples{
getValidPlots(mlr.benchmark.example)
}
