% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvJsonHandler.R
\name{jsonExport}
\alias{jsonExport}
\title{Export a benchmarkVis data table as json file}
\usage{
jsonExport(dt, file.path)
}
\arguments{
\item{dt}{the benchmarkVis data table}

\item{file.path}{path to save the file to}
}
\description{
Export the specified benchmarkVis compatible data table as an json file. Lists will be converted into a String:
"list(valA = 123, valB = 'xyz')".
}
\examples{
jsonExport(mlr.benchmark.example, "test.json")
}
