% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvJsonHandler.R
\name{jsonImport}
\alias{jsonImport}
\title{Import a json file as a benchmarkVis data table}
\usage{
jsonImport(input.json)
}
\arguments{
\item{input.json}{path to the input json file}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Import a matching json file as a benchmarkVis compatible data table.
Lists must be a String with following structure:
"list(valA = 123, valB = 'xyz')".
Check the file "extdata/ml.benchmark2.json" to see an example of the correct file structure
}
