% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{large.benchmark}
\alias{large.benchmark}
\title{Large benchmark}
\description{
A large example of a benchmark created with mlr. It was converted to the benchmarkVis table structure by the mlrBenchmarkWrapper.
The table consists of 18 columns and 210 rows.
Problems: iris, sonar, pid, ring, wave, breast.cancer.
Algorithms: lda, rpart, xgboost, randomForest, naiveBayes, ksvm, knn.
Replications: Subsample, Holdout, CV 5, LOO, Bootstrap.
The file is accessible through "large.benchmark".
}
\references{
See \url{data/large.benchmark.RData}
}
\keyword{data}
