% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{listPlots}
\alias{listPlots}
\title{Get list of all possible plots}
\usage{
listPlots()
}
\value{
vector containing all possbile plots
}
\description{
Get a complete list of all plots implemented within the benchmarkVis application
}
