% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/information.R
\name{listWrappers}
\alias{listWrappers}
\title{Get list of all possible wrappers}
\usage{
listWrappers()
}
\value{
vector containing all possbile wrappers
}
\description{
Get a complete list of all wrappers implemented within the benchmarkVis application
}
\examples{
listWrappers()
}
