% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrBenchmarkWrapper.R
\name{makeMlrBenchmarkWrapper}
\alias{makeMlrBenchmarkWrapper}
\title{Insert mlr benchmark object into benchmarkVis application}
\usage{
makeMlrBenchmarkWrapper(input.file)
}
\arguments{
\item{input.file}{Path to the input mlr benchmark file}
}
\value{
a dataframe with the benchmarkVis specific structure
}
\description{
Create a dataframe useable within the benchmarkVis application out of an mlr benchmark object. All importont imformation will be read out of the input file and transformed in a matching build dataframe
}
