% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{microbenchmark.example}
\alias{microbenchmark.example}
\title{microbenchmark input}
\description{
Contains the wrapped microbenchmark result (\code{microbenchmark.example}).
It is a basic Version of a useable benchmark result dataframe.
It was converted to a dataframe userable within the benchmarkVis application through the microbenchmarkWrapper.
The file is accessible through "microbenchmark.example".
}
\references{
See \url{data/microbenchmark.example.RData}
}
\keyword{data}
