% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ml.example}
\alias{ml.example}
\title{Example input dataframe}
\description{
Contains the dataframe (\code{ml.example}).
This file is a basic Version of a useable benchmark result dataframe.
It can be loaded with "data("mlExample")" and will then be accessible through "ml.df".
}
\references{
See \url{data/ml.example.RData}
}
\keyword{data}
