% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ml.example2}
\alias{ml.example2}
\title{Example Json file}
\description{
Contains the same data table as (\code{mlr.benchmark.example}).
This file is an example of how a useable json file should be structured.
It can be loaded into the benchmarkVis application through the jsonImport method.
}
\references{
See \url{extdata/ml.example2.json}
}
\keyword{data}
