% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mlr.benchmark.example}
\alias{mlr.benchmark.example}
\title{mlr benchmark input}
\description{
Contains the wrapped mlr benchmark result (\code{mlr.benchmark.example}).
It is a basic Version of a useable mlr benchmark result dataframe.
It was converted to a dataframe useable within the benchmarkVis application through the mlrBenchmarkWrapper.
The file is accessible through "mlr.benchmark.example".
}
\references{
See \url{data/mlr.benchmark.example.RData}
}
\keyword{data}
