% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mlr.benchmark}
\alias{mlr.benchmark}
\title{mlr benchmark result}
\description{
Contains the mlr benchmark result (\code{mlr.benchmark})
This file an example of a mlr benchmark result export.
It can be converted to a dataframe userable within the benchmarkVis application through the mlrBenchmarkWrapper.
The file can be loaded with "data("mlrBenchmark")" and will then be accessible through "mlr.bmr".
Afterwards it should be inserted into the makeMlrBenchmarkWrapper() method.
}
\references{
See \url{data/mlr.benchmark.RData}
}
\keyword{data}
