% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mlr.tuning.example}
\alias{mlr.tuning.example}
\title{mlr tuning input}
\description{
Contains the wrapped mlr tuning result (\code{mlr.tuning.example}).
It is a basic Version of a useable mlr tuning result data table.
It was converted to a data table useable within the benchmarkVis application through the mlrTuningWrapper.
The file is accessible through "mlr.tuning.example".
}
\references{
See \url{data/mlr.tuning.example.RData}
}
\keyword{data}
