% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{mlrBenchmark}
\alias{mlrBenchmark}
\title{This file an example of an mlr benchmark result export.
It can be converted to a dataframe userable within the benchmarkVis application through the mlrBenchmarkWrapper.
The file can be loaded with "readRDS("./data/mlrBenchmark.rds")"
Afterwards it should be inserted into the makeMlrBenchmarkWrapper() method.}
\description{
This file an example of an mlr benchmark result export.
It can be converted to a dataframe userable within the benchmarkVis application through the mlrBenchmarkWrapper.
The file can be loaded with "readRDS("./data/mlrBenchmark.rds")"
Afterwards it should be inserted into the makeMlrBenchmarkWrapper() method.
}
\references{
See \url{data/mlrBenchmark.rds}
}
\keyword{data}
