% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAggregation.R
\name{parser.agg.input}
\alias{parser.agg.input}
\title{parse the text input of agg function.}
\usage{
parser.agg.input(aggfun)
}
\arguments{
\item{aggfun}{the text include agg function.}
}
\value{
a vector
}
\description{
parse the text include agg function to a vector of function names.
}
\examples{
parser.agg.input("mean,standard deviation,median")
}
