% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{parser.function.list}
\alias{parser.function.list}
\title{parse the text input of a list of functions}
\usage{
parser.function.list(function.list)
}
\arguments{
\item{function.list}{list of functions to parse}
}
\value{
a vector
}
\description{
parse the list of functions to a vector of function names.
}
