% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsHandler.R
\name{rdsExport}
\alias{rdsExport}
\title{Export a benchmarkVis data table as .rds file}
\usage{
rdsExport(dt, file.path)
}
\arguments{
\item{dt}{the benchmarkVis data table}

\item{file.path}{path to save the file to}
}
\description{
Export the specified benchmarkVis compatible data table as an .rds file.
}
\examples{
rdsExport(mlr.benchmark.example, "test.rds")
}
