% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdsHandler.R
\name{rdsImport}
\alias{rdsImport}
\title{Import an .rds file as a benchmarkVis data table}
\usage{
rdsImport(input.rds)
}
\arguments{
\item{input.rds}{path to the input .rds file}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Import a matching .rds file as a benchmarkVis compatible data table.
Check the file "extdata/ml.benchmark.rds" to see an example of the correct file structure
}
