% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvJsonHandler.R
\name{tableTransformationExport}
\alias{tableTransformationExport}
\title{Transform table for export}
\usage{
tableTransformationExport(dt, is.csv)
}
\arguments{
\item{dt}{the benchmarkVis compatible data table}

\item{is.csv}{working with csv or json file}
}
\value{
transformed table
}
\description{
Do all the table tranformations needed for file export. Convert columns from lists to strings.
For csv & json: Transform parameter lists
For csv only: Transform vectors
}
