% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvJsonHandler.R
\name{tableTransformationImport}
\alias{tableTransformationImport}
\title{Transform table for import}
\usage{
tableTransformationImport(dt, is.csv)
}
\arguments{
\item{dt}{the not yet benchmarkVis compatible data table}

\item{is.csv}{working with csv or json file}
}
\value{
transformed table
}
\description{
Do all the table tranformations needed for file import. Convert columns from string to lists.
For csv & json: Transform parameter lists
For csv only: Transform vectors
}
