% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{transformation.apply}
\alias{transformation.apply}
\title{apply transformation functions}
\usage{
transformation.apply(original.data, columns.to.transform,
  transformation.functions)
}
\arguments{
\item{original.data}{a data table in benchmarkVis format}

\item{columns.to.transform}{the list of columns to apply the functions on}

\item{transformation.functions}{the functions for transofrmation of data}
}
\value{
a data table containing the original data and transformed versions of columns
}
\description{
do transformation and return a new data table
}
\examples{
transformation.apply(original.data = mlr.benchmark.example, columns.to.transform = c("measure.mmce.test.mean", "measure.ber.test.mean"), transformation.functions = c("log2"))
}
