% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{unprettifyPlotName}
\alias{unprettifyPlotName}
\title{Get name of a plot in the benchmarkVis format: "createTypeName"}
\usage{
unprettifyPlotName(plot.name)
}
\arguments{
\item{plot.name}{the plot you want to unbeautify}
}
\value{
name of the plot in the benchmarkVis format
}
\description{
Get the benchmarkVis name version of a plot.
}
