% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbenchmarkWrapper.R
\name{useMicrobenchmarkFileWrapper}
\alias{useMicrobenchmarkFileWrapper}
\title{Insert mlr benchmark RDS file into benchmarkVis application}
\usage{
useMicrobenchmarkFileWrapper(input.file)
}
\arguments{
\item{input.file}{Path to the input microbenchmark RDS file}
}
\value{
a dataframe with the benchmarkVis specific structure
}
\description{
Load the specified file and pass it on the the useMicrobenchmarkWrapper function.
Create a dataframe useable within the benchmarkVis application out of an microbenchmark object.
All important imformation will be exluded from the input object and transformed into a appropriate dataframe
}
