% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrBenchmarkWrapper.R
\name{useMlrBenchmarkWrapperWithRdsFile}
\alias{useMlrBenchmarkWrapperWithRdsFile}
\title{Insert mlr benchmark RDS file into benchmarkVis application}
\usage{
useMlrBenchmarkWrapperWithRdsFile(input.file)
}
\arguments{
\item{input.file}{Path to the input mlr benchmark RDS file}
}
\value{
a dataframe with the benchmarkVis specific structure
}
\description{
Load the specified file and pass it on the the useMlrBenchmarkWrapper function.
Create a dataframe useable within the benchmarkVis application out of an mlr benchmark object.
All importont imformation will be read out of the input file and transformed in a matching build dataframe
}
