% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTuningWrapper.R
\name{useMlrTuningFileWrapper}
\alias{useMlrTuningFileWrapper}
\title{Insert mlr tuning RDS file into benchmarkVis application}
\usage{
useMlrTuningFileWrapper(input.file)
}
\arguments{
\item{input.file}{Path to the input mlr tuning RDS file}
}
\value{
a dataframe with the benchmarkVis specific structure
}
\description{
Load the specified file and pass it on the the useMlrTuningWrapper function.
Create a dataframe useable within the benchmarkVis application out of an mlr tuning object.
All important imformation will be exluded from the input object and transformed into a appropriate dataframe
}
