% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrTuningWrapper.R
\name{useMlrTuningWrapper}
\alias{useMlrTuningWrapper}
\title{Insert mlr tuning object into benchmarkVis application}
\usage{
useMlrTuningWrapper(res)
}
\arguments{
\item{res}{a mlr tuning result object}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Create a data table useable with the benchmarkVis application out of an mlr tuning object.
All important imformation will be exluded from the input object and transformed into a appropriate data table
}
\examples{
library(mlr)
num.ps = makeParamSet(
 makeNumericParam("C", lower = -10, upper = 10, trafo = function(x) 10 ^ x),
 makeNumericParam("sigma", lower = -10, upper = 10, trafo = function(x) 10 ^ x)
)
ctrl = makeTuneControlRandom(maxit = 100L)
rdesc = makeResampleDesc("CV", iters = 3L)
res = tuneParams("classif.ksvm", task = iris.task, resampling = rdesc, par.set = num.ps, control = ctrl,
 measures = list(acc, setAggregation(acc, test.sd)), show.info = FALSE)
dt = useMlrTuningWrapper(res)
}
