% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbenchmarkWrapper.R
\name{useRbenchmarkFileWrapper}
\alias{useRbenchmarkFileWrapper}
\title{Insert rbenchmark RDS file into benchmarkVis application}
\usage{
useRbenchmarkFileWrapper(input.file)
}
\arguments{
\item{input.file}{Path to the input rbenchmark RDS file}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Load the specified file and pass it on the the useRbenchmarkWrapper function.
Create a data table useable within the benchmarkVis application out of an rbenchmark object.
All important imformation will be exluded from the input object and transformed into a appropriate data table
}
