% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbenchmarkWrapper.R
\name{useRbenchmarkWrapper}
\alias{useRbenchmarkWrapper}
\title{Insert rbenchmark object into benchmarkVis application}
\usage{
useRbenchmarkWrapper(benchmark)
}
\arguments{
\item{benchmark}{a rbenchmark object}
}
\value{
a dataframe with the benchmarkVis specific structure
}
\description{
Create a dataframe useable within the benchmarkVis application out of an rbenchmark object.
All important imformation will be exluded from the input object and transformed into a appropriate dataframe
}
\examples{
library(rbenchmark)
x = runif(1000)
benchmark = rbenchmark::benchmark(
 shell_sort = sort(x, method = "shell"),
 quick_sort = sort(x, method = "quick"),
 radix_sort = sort(x, method = "radix"),
 columns = c(
   "test", "replications", "elapsed"),
 order = "elapsed",
 replications = c(100, 20)
)
df = useRbenchmarkWrapper(benchmark)
}
