% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doAggregation.R
\name{check.agg.valid}
\alias{check.agg.valid}
\title{check the input is a valid aggragate function name}
\usage{
check.agg.valid(x)
}
\arguments{
\item{x}{the string of the aggragate function.}
}
\value{
bool
}
\description{
check the input is a valid aggragate function name
}
\examples{
check.agg.valid("mean")
}
