% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkStructure.R
\name{checkStructure}
\alias{checkStructure}
\title{Check the structure of a benchmarkVis compatible dataframe}
\usage{
checkStructure(df)
}
\arguments{
\item{df}{dataframe to check}
}
\value{
True if test was successful
}
\description{
Check if the structure of a dataframe satisfies the benchmarkVis requirements.
If it does the method will return TRUE else an error will be thrown
}
\examples{
checkStructure(microbenchmark.example)
}
