% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barPlot.R
\name{createBarPlot}
\alias{createBarPlot}
\title{Create a bar plot}
\usage{
createBarPlot(dt, measure, color.by = "algorithm", group.color = FALSE,
  stacked = FALSE)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the column name containing the results of a measure}

\item{color.by}{the column to color the bars with. Possibilities: "algorithm", "problem", "replication" (default: "algorithm")}

\item{group.color}{instead of coloring the column specified by color.by it will be grouped. Color will be chosen from the remaining main columns instead (default: FALSE)}

\item{stacked}{defines if the grouped bars should be stacked. Just working with "group.color" = TRUE (default: FALSE)}
}
\value{
a plotly bar plot
}
\description{
Create a plotly bar plot out of a benchmarkVis compatible data table.
The created bar plot shows the measure result for each benchmark entry.
x-Axis: the measure.
y-Axis: the benchmark entries.
}
\examples{
createBarPlot(mlr.benchmark.example, "measure.mmce.test.mean")
}
