% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BoxPlot.R
\name{createBoxPlot}
\alias{createBoxPlot}
\title{Create a box plot}
\usage{
createBoxPlot(df, measure)
}
\arguments{
\item{df}{compatible dataframe}

\item{measure}{measure for comparison}
}
\value{
a box plot
}
\description{
Create a plotly box plot out of a benchmarkVis compatible dataframe.
The created box plot allows for comparison of algorithms based on a given performance measure for different datasets
}
\examples{
createBoxPlot(mlr.benchmark.example, 'mmce.test.mean')
}
