% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/critDifferencesPlot.R
\name{createCritDifferencesPlot}
\alias{createCritDifferencesPlot}
\title{create a crit differences plot}
\usage{
createCritDifferencesPlot(dt, measure, group.by = "problem")
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the measure to plot}

\item{group.by}{the column to group the markers by. Possibilities: "algorithm", "problem", "replication" (default: "problem")}
}
\value{
a crit differences plot
}
\description{
Create critDifferences plot out of a benchmarkVis compatible data table.
}
\examples{
createCritDifferencesPlot(mlr.benchmark.example, "measure.mmce.test.mean")
}
