% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualMeasurePlot.R
\name{createDualMeasurePlot}
\alias{createDualMeasurePlot}
\title{Create a Dual Measure plot}
\usage{
createDualMeasurePlot(dt, measure1, measure2, color.by = "algorithm",
  interaction.column = "none", pointsize = 4L, jitter = 0)
}
\arguments{
\item{dt}{compatible data table}

\item{measure1}{the first measure for comparison}

\item{measure2}{the second measure  for comparison}

\item{color.by}{the column to color the markers with. Possibilities: "algorithm", "problem", "replication" (default: "algorithm")}

\item{interaction.column}{will result in colors being an interaction of color.by and this value. Possibilities: "algorithm", "problem", "replication", "none" (default: "none")}

\item{pointsize}{the size of ploted point}

\item{jitter}{small vertical jitter to deal with overplotting in case of equal scores}
}
\value{
a dual measure plot
}
\description{
Create a Dual Measure plot out of a benchmarkVis compatible data table.
The created Dual Measure plot allows to compare the performance of the color.by input based on two giving performance measures.
x-Axis: measure1.
y-Axis: measure2.
}
\examples{
createDualMeasurePlot(mlr.benchmark.example, 'measure.mmce.test.mean','measure.ber.test.mean')
}
