% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listPlots.R
\name{createListLinePlot}
\alias{createListLinePlot}
\title{Create a list line plot}
\usage{
createListLinePlot(dt, list.measure, cumulative.function = "id")
}
\arguments{
\item{dt}{campatible data table}

\item{list.measure}{the column name containing the list of a specific measure}

\item{cumulative.function}{the cumulative function to use for the list measure (default: "id")}
}
\value{
a plotly line plot
}
\description{
Create a plotly line plot out of a benchmarkVis compatible data table.
The created line chart shows the change within the specified list measure for each iteration
A cumulative function can be defined to get a better feeling of the course of the line chart.
Possible parameters for the cumulative function are: "id" (identity function), "min", "max", "mean" - default is id.
}
\examples{
createListLinePlot(microbenchmark.example, "list.values")
}
