% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallelCoordinatesPlot.R
\name{createParallelCoordinatesPlot}
\alias{createParallelCoordinatesPlot}
\title{Create a parallel coordinates plot}
\usage{
createParallelCoordinatesPlot(dt)
}
\arguments{
\item{dt}{compatible data table}
}
\value{
a plotly parallel coordinates plot
}
\description{
Create a plotly parallel coordinates plot out of a benchmarkVis compatible data table.
The created parallel coordinates plot shows the relationship between problem, algorithm and all measures.
}
\examples{
createParallelCoordinatesPlot(mlr.benchmark.example)
}
