% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPlot.R
\name{createRankPlot}
\alias{createRankPlot}
\title{Create a rank plot}
\usage{
createRankPlot(df, measure, problem, algorithm)
}
\arguments{
\item{df}{compatible dataframe}

\item{measure}{measure for Ranking}
}
\value{
a rank plot
}
\description{
Create a plotly rank plot out of a benchmarkVis compatible dataframe.
The created rank plot allows for Ranking of algorithms based on a given performance measure for different datasets
}
\examples{
createRankPlot(mlr.benchmark.example, 'mmce.test.mean')
}
