% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RankPlot.R
\name{createRankPlot}
\alias{createRankPlot}
\title{Create a rank plot}
\usage{
createRankPlot(dt, measure)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{measure for Ranking}
}
\value{
a rank plot
}
\description{
Create a plotly rank plot out of a benchmarkVis compatible data table.
The created rank plot allows for Ranking of algorithms based on a given performance measure for different datasets
}
\examples{
createRankPlot(mlr.benchmark.example, 'measure.mmce.test.mean')
}
