% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicationPlots.R
\name{createReplicationLinePlot}
\alias{createReplicationLinePlot}
\title{Create a replication line plot}
\usage{
createReplicationLinePlot(df, replication.measure)
}
\arguments{
\item{df}{campatible dataframe}

\item{replication.measure}{the column name containing the replications of a specific measure}
}
\value{
a plotly line plot
}
\description{
Create a plotly line plot out of a benchmarkVis compatible dataframe.
The created line chart shows the change within the specified measure for each iteration of the replication
}
\examples{
createReplicationLinePlot(microbenchmark.example, "replication.values")
}
