% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{createScatterPlot}
\alias{createScatterPlot}
\title{Create a scatter plot}
\usage{
createScatterPlot(dt, measure1, measure2, color.algorithm.only = TRUE,
  pointsize = 4L, jitter = 1)
}
\arguments{
\item{dt}{compatible data table}

\item{measure1}{the first measure for comparison}

\item{measure2}{the second measure  for comparison}

\item{pointsize}{the size of ploted point}

\item{jitter}{small vertical jitter to deal with overplotting in case of equal scores}
}
\value{
a scatter plot
}
\description{
Create a scatter plot out of a benchmarkVis compatible data table.
The created scatter plot allows to compare the performance of interaction(algorithm and task) based on two giving performance measures for different datasets.
x-Axis: measure1.
y-Axis: measure2.
color: the interaction of algorithm and problem.
}
\examples{
createScatterPlot(mlr.benchmark.example, 'measure.mmce.test.mean','measure.ber.test.mean')
}
