% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatterPlot.R
\name{createScatterPlot}
\alias{createScatterPlot}
\title{Create a scatter plot}
\usage{
createScatterPlot(dt, measure)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{the measure to plot}
}
\value{
a scatter plot
}
\description{
Create a scatter plot out of a benchmarkVis compatible data table.
The created scatter plot allows to compare the performances different algorithms based on the giving performance measure.
x-Axis: problems.
y-Axis: measure.
color: the algorithm.
}
\examples{
createScatterPlot(mlr.benchmark.example, 'measure.mmce.test.mean')
}
