% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaryPlot.R
\name{createSummaryPlot}
\alias{createSummaryPlot}
\title{Create a summary plot}
\usage{
createSummaryPlot(dt, measure)
}
\arguments{
\item{dt}{compatible data table}

\item{measure}{measure for comparison}
}
\value{
a summary plot
}
\description{
Create a partially filled bar charts plot out of a benchmarkVis compatible data table.
each line refers to a task, On that line the aggregated scores for all learners are plotted.
The created partially bar charts allow to compare the performance of every combination of task and learner
x-Axis: the measure.
y-Axis: the problem.
color: the algorithem.
}
\examples{
createSummaryPlot(mlr.benchmark.example, 'measure.mmce.test.mean')
}
