% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvHandler.R
\name{csvExport}
\alias{csvExport}
\title{Export a benchmarkVis data frame as csv file}
\usage{
csvExport(df, file.path)
}
\arguments{
\item{df}{the benchmarkVis data frame}

\item{file.path}{path to save the file to}
}
\description{
Export the specified benchmarkVis compatible data frame as an csv file. Lists will be converted into a String:
"list(valA = abc, valB = xyz)"
}
\examples{
csvExport(mlr.benchmark.example, "test.csv")
}
