% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csvHandler.R
\name{csvImport}
\alias{csvImport}
\title{Import a csv file as a benchmarkVis data table}
\usage{
csvImport(input.csv)
}
\arguments{
\item{input.csv}{path to the input csv file}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Import a matching csv file as a benchmarkVis compatible data table
Check the file "data/ml.benchmark.csv" to see an example of the correct file structure
}
