% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{get.agg.result}
\alias{get.agg.result}
\title{get aggregation result}
\usage{
get.agg.result(fun, prefix_str, groupby, aggcol, df)
}
\arguments{
\item{fun}{aggregate functions}

\item{prefix_str}{the column name's prefix}

\item{groupby}{the list of columns name that will be grouped}

\item{aggcol}{the list of columns name that will be aggregated}

\item{df}{the input dataframe}
}
\value{
a dataframe with aggregated column
}
\description{
do aggregation and return a new data table
}
