% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{get.num.columns.name}
\alias{get.num.columns.name}
\title{do aggregation result}
\usage{
get.num.columns.name(data)
}
\arguments{
\item{data}{a dataframe}
}
\value{
list of the nummeric column names
}
\description{
do aggregation and return a new data table.
the aggcol must include problem and algorithm
}
