% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ml.example}
\alias{ml.example}
\title{Example CSV file}
\description{
Contains the same data table as (\code{mlr.benchmark.example}).
This file is an example of how a useable csv file should be structured.
It can be loaded into the benchmarkVis application through the csvImport method.
}
\references{
See \url{data/ml.example.csv}
}
\keyword{data}
