% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rbenchmark.example}
\alias{rbenchmark.example}
\title{rbenchmark input}
\description{
Contains the wrapped rbenchmark result (\code{rbenchmark.example}).
It is a basic Version of a useable benchmark result dataframe.
It was converted to a dataframe useable within the benchmarkVis application through the rbenchmarkWrapper.
The file is accessible through "rbenchmark.example".
}
\references{
See \url{data/rbenchmark.example.RData}
}
\keyword{data}
