% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{transformation.apply}
\alias{transformation.apply}
\title{get transformation result}
\usage{
transformation.apply(original.data, columns.to.transform,
  transformation.functions)
}
\arguments{
\item{groupby}{the list of columns name that will be grouped}

\item{aggcol}{the list of columns name that will be aggregated}

\item{aggfun}{the function to aggregate with}

\item{df}{the input dataframe}
}
\value{
a dataframe
}
\description{
do transformation and return a new data table.
the groupby columns must include problem and algorithm
}
\examples{
transformation.apply(original.data = mlr.benchmark.example, columns.to.transform = c("measure.mmce.test.mean", "measure.ber.test.mean"), transformation.functions = c("log2"))
}
