% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableOps.R
\name{transformation.apply}
\alias{transformation.apply}
\title{get transformation result}
\usage{
transformation.apply()
}
\arguments{
\item{groupby}{the list of columns name that will be grouped}

\item{aggcol}{the list of columns name that will be aggregated}

\item{aggfun}{the function to aggregate with}

\item{df}{the input dataframe}
}
\value{
a dataframe
}
\description{
do transformation and return a new data table.
the groupby columns must include problem and algorithm
}
\examples{
transformation.apply(groupby= c("problem", "algorithm"), aggfun= c("mean"), aggcol= c("measure.mmce.test.mean", "measure.ber.test.mean"), df= mlr.benchmark.example)
}
