% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microbenchmarkWrapper.R
\name{useMicrobenchmarkWrapper}
\alias{useMicrobenchmarkWrapper}
\title{Insert microbenchmark object into benchmarkVis application}
\usage{
useMicrobenchmarkWrapper(benchmark)
}
\arguments{
\item{benchmark}{a microbenchmark object}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Create a data table useable with the benchmarkVis application out of an microbenchmark object.
All important imformation will be exluded from the input object and transformed into a appropriate data table
}
\examples{
library(microbenchmark)
x = runif(100)
benchmark = microbenchmark(
sqrt(x),
x ^ 0.5
)
dt = useMicrobenchmarkWrapper(benchmark)
}
