% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlrBenchmarkWrapper.R
\name{useMlrBenchmarkWrapper}
\alias{useMlrBenchmarkWrapper}
\title{Insert mlr benchmark object into benchmarkVis application}
\usage{
useMlrBenchmarkWrapper(bmr)
}
\arguments{
\item{bmr}{a mlr benchmark result object}
}
\value{
a data table with the benchmarkVis specific structure
}
\description{
Create a data table useable with the benchmarkVis application out of an mlr benchmark object.
All important imformation will be exluded from the input object and transformed into a appropriate data table
}
\examples{
library(mlr)
lrns = list(makeLearner("classif.lda"), makeLearner("classif.rpart"))
rdesc = makeResampleDesc("Holdout")
bmr = benchmark(lrns, sonar.task, rdesc)
dt = useMlrBenchmarkWrapper(bmr)
}
